SpreadsheetFormatterFetcherWatcher.onSpreadsheetFormatterSelectorMenu SpreadsheetFormatterSelectorMenuList was SpreadsheetFormatterSelectorMenu
