+++ src/main/java/walkingkooka/spreadsheet/dominokit/net/SpreadsheetFormatterFetcherWatchersMenuListEvent.java
@@ -1,0 +1,64 @@
+/*
+ * Copyright 2023 Miroslav Pokorny (github.com/mP1)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package walkingkooka.spreadsheet.dominokit.net;
+
+import walkingkooka.spreadsheet.SpreadsheetId;
+import walkingkooka.spreadsheet.dominokit.AppContext;
+import walkingkooka.spreadsheet.server.formatter.SpreadsheetFormatterSelectorMenuList;
+
+/**
+ * The event payload used by {@link SpreadsheetFormatterFetcherWatchers} when it receives a {@link SpreadsheetFormatterSelectorMenuList}.
+ */
+final class SpreadsheetFormatterFetcherWatchersMenuListEvent extends FetcherWatchersEvent<SpreadsheetFormatterFetcherWatcher> {
+
+    static SpreadsheetFormatterFetcherWatchersMenuListEvent with(final SpreadsheetId id,
+                                                                 final SpreadsheetFormatterSelectorMenuList menuList,
+                                                                 final AppContext context) {
+        return new SpreadsheetFormatterFetcherWatchersMenuListEvent(
+                id,
+                menuList,
+                context
+        );
+    }
+
+    private SpreadsheetFormatterFetcherWatchersMenuListEvent(final SpreadsheetId id,
+                                                             final SpreadsheetFormatterSelectorMenuList menuList,
+                                                             final AppContext context) {
+        super(context);
+        this.id = id;
+        this.menuList = menuList;
+    }
+
+    @Override
+    void fire(final SpreadsheetFormatterFetcherWatcher watcher) {
+        watcher.onSpreadsheetFormatterSelectorMenuList(
+                this.id,
+                this.menuList,
+                this.context
+        );
+    }
+
+    private final SpreadsheetId id;
+
+    private final SpreadsheetFormatterSelectorMenuList menuList;
+
+    @Override
+    public String toString() {
+        return this.id + " " + this.menuList + " " + this.context;
+    }
+}
